First Demo
The demo from this commit can display buttons in the style of game posters, which contain a background image, sans-logo, and a dynamic floating logo that will move to the center once the button is hovered over. These buttons can be created through the form on the top of the page.